Updated notes 2 to enhance the TODO for meet 3
I went ahead and updated the to-do list right now, so we don't have to spend time figuring out the specific goals at the beginning of the meet. I realize this is an extensive todo, so it might extend into next week. I would like to get some of it done, however.

I might also not be able to make it to this meet. So, I apologize in advance.